refactor(daemon): break the memory <-> memory-plugin package cycle
memory-plugin/outbox.ts imported canonicalAgentMemoryKey from memory/recall.js
while memory/ imports memory-plugin, a directory-level cycle. Move the pure key
normalization into a new leaf memory/keys.ts that imports nothing, and point
recall.ts, provider.ts, outbox.ts and the test at it. No re-export left behind.
